+++ src/pages/sitemap.xml.ts
@@ -1,0 +1,13 @@
+import { allSeoPaths, canonicalUrl } from "../data/site";
+
+export function GET() {
+  const urls = allSeoPaths()
+    .map((path) => `  <url><loc>${canonicalUrl(path)}</loc></url>`)
+    .join("\n");
+
+  return new Response(`<?xml version="1.0" encoding="UTF-8"?>\n<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">\n${urls}\n</urlset>\n`, {
+    headers: {
+      "Content-Type": "application/xml; charset=utf-8",
+    },
+  });
+}